Unified ios and jvm message generation

+++ grpc-multiplatform-lib/src/androidJvmCommon/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/message/KMMessage.kt
@@ -8,7 +8,7 @@ actual interface KMMessage {
     val requiredSize: Int
 
     fun serialize(): ByteArray {
-        val data = ByteArray(requiredSize.toInt())
+        val data = ByteArray(requiredSize)
         val stream = com.google.protobuf.CodedOutputStream.newInstance(ByteBuffer.wrap(data))
         serialize(CodedOutputStream(stream))
